move conditional attention to trm/layers
